Mark missing LTR icons
Some LTR icons that were missed in commit
c5fb9b7823dc99781f6e39964b68f5ec94de2a24 are now categorized as such. A
few wrongly marked LTR icons are now reverted to relative direction as
well.

Additional icons are now also marked as LTR, based on the Material
design spec guidelines posted at:
https://www.google.com/design/spec/usability/bidirectionality.html#bidirectionality-ui-mirroring-overview
Specifically, the following previously unobserved rules are now
observed:

 - Physical objects are not mirrored.
 - Icons that are depict objects in such a direction would result from
   being held in the right hand are not considered directional.
 - Media player icons are not mirrored.
 - Icons that unambiguously represent clocks or clockwise motion are not
   mirrored. However, icons that can represent either the linear or
   circular direction of time are mirrored.

Composite icons should usually be marked conservatively, unless the
combination results in something that should be mirrored. If the
composite icon base was part of a set that is likely to be used
interchangeably or in parallel, then the conservative ruling is applied
to all of them.

Note that these rules were applied to the current state of the icon, not
directly on the type of the represented object itself. Therefore, many
items that should be marked as LTR are not, since they are completely
symmetrical and are drawn exactly the same from both directions. This is
both to save effort, and because there is no way to be sure beforehand
of the exact nature of the icon that may be used to represent any given
object, thus leaving room for ambiguity. However, icons of the same type
that are grouped together in the naming scheme may be all be defined as
unmirrorable, even if some of them are symmetrical, unless they are
liable to have overlays or combinations with other icons.