Use new layout direction API in IconDrawable
Marshmallow made the methods for setting and querying drawable layout
directions part of the public API, as well as a callback method on value
change. This is now used for the RTL auto-mirroring implementation in
IconDrawable in Marshmallow and above, instead of the previous hack
approach to attempt to cast the callback to a View in order to query it
for it's layout direction. The compile SDK version has been incremented
to 23 to enable usage of the new API. New text direction properties that
were introduced in Marshmallow are now also handled for text views in
CustomTypefaceSpan.